requirements.txt
acres==0.3.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.3.0
blinker==1.9.0
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
configobj==5.0.9
configparser==7.2.0
contourpy==1.3.2
cycler==0.12.1
datasets==3.5.0
dill==0.3.8
etelemetry==0.3.1
filelock==3.18.0
fitz==0.0.1.dev2
Flask==3.1.0
fonttools==4.57.0
frozenlist==1.6.0
fsspec==2024.12.0
httplib2==0.22.0
huggingface-hub==0.30.2
idna==3.10
imageio==2.37.0
importlib_resources==6.5.2
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
lazy_loader==0.4
looseversion==1.3.0
lxml==5.3.2
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
networkx==3.4.2
nibabel==5.3.2
nipype==1.10.0
numpy==2.2.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pdf2image==1.17.0
pillow==11.2.1
propcache==0.3.1
prov==2.0.1
puremagic==1.28
pyarrow==19.0.1
pydot==3.0.4
PyMuPDF==1.25.5
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==6.3.2
regex==2024.11.6
reportlab==4.4.0
requests==2.32.3
safetensors==0.5.3
scikit-image==0.25.2
scipy==1.15.2
shapely==2.1.0
simplejson==3.20.1
six==1.17.0
sympy==1.13.1
tifffile==2025.3.30
timm==1.0.15
tokenizers==0.21.1
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
traits==7.0.2
transformers==4.51.3
triton==3.2.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
xxhash==3.5.0
yarl==1.20.0
